dsa: move `Signature` to toplevel
It was previously in a `sig` module to avoid clashing with the
`signature` crate re-export.

This structure matches the one used by the `ecdsa` crate.
